requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
pandas
numpy
joblib
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
